removed field tag from Token. It's never used with any sensible purpose.
In Chunking, the only time is used, it's to set the sample.label. Since the tag is never changed anywhere except at creation, is equal to "". Therefore the deleted line is equivalent to set the label to "". And since the creation of a ME_Sample makes the label be set to "" by default, the line is unnecessary